requirements.txt
djangorestframework
django-cors-headers
django-import-export
django-filter
pillow
gunicorn
whitenoise
dj-database-url
psycopg2-binary
